+++ winsup/cygwin/include/sys/dirent.h
@@ -42,7 +42,7 @@ typedef struct __DIR
   char *__d_dirname;			/* directory name with trailing '*' */
   __int32_t __d_position;			/* used by telldir/seekdir */
   int __d_fd;
-  unsigned __d_internal;
+  uintptr_t __d_internal;
   void *__handle;
   void *__fh;
   unsigned __flags;
